Create the rda package for a plot or table

Create the rda package for a plot or table

create_latex_table(data, caption, label)

Arguments

data

Data frame. A dataframe-based table

caption

String. Table caption

label

String. Table label

Value

A table based in LaTeX.

Examples

create_latex_table(
  data = as.data.frame(head(mtcars, 6)),
  caption = "My caption",
  label = "My label"
)
#> [1] "\\tagpdfsetup{table/header-rows={1}}\n\\begin{center}\n\\captionof{table}{My caption}\n\\label{My label}\n\n\\begin{tabular}{lrrrrrrrrrrr}\n\\toprule\n  & mpg & cyl & disp & hp & drat & wt & qsec & vs & am & gear & carb\\\\\n\\midrule\nMazda RX4 & 21.0 & 6 & 160 & 110 & 3.90 & 2.620 & 16.46 & 0 & 1 & 4 & 4\\\\\nMazda RX4 Wag & 21.0 & 6 & 160 & 110 & 3.90 & 2.875 & 17.02 & 0 & 1 & 4 & 4\\\\\nDatsun 710 & 22.8 & 4 & 108 & 93 & 3.85 & 2.320 & 18.61 & 1 & 1 & 4 & 1\\\\\nHornet 4 Drive & 21.4 & 6 & 258 & 110 & 3.08 & 3.215 & 19.44 & 1 & 0 & 3 & 1\\\\\nHornet Sportabout & 18.7 & 8 & 360 & 175 & 3.15 & 3.440 & 17.02 & 0 & 0 & 3 & 2\\\\\nValiant & 18.1 & 6 & 225 & 105 & 2.76 & 3.460 & 20.22 & 1 & 0 & 3 & 1\\\\\n\\bottomrule\n\\end{tabular}\n\\end{center}"