SS3 Example data

Included data set that represents a Report.sso file converted using convert_output(). This example is from the 2022 Petrale sole stock assessment.

example_data

Format

A tibble with 591109 rows and 33 variables:

label

Standard name for estimate or value name

estimate

Actual value of the label

year

Annual indexing value

fleet

Names of fleets or surveys indexed by the data

sex

Native reference to male, female, unknown, or none

area

Specified areas by the model

growth_pattern

Indexing column of data

uncertainty

Value of uncertainty associated with the label column

uncertainty_label

Uncertainty label or name associated with the label columns

module_name

Name of keyword (SS3) or list indexing in the original model output for tracking purposes

time

Time sometimes referenced in decimals to the year and month

era

"time" for current time series in model; "fore" representing the projected or forecasted years of the model

month

Month factor

season

Season usually associated with year

subseason

Subseason when used

birthseas

Birthseason found in SS3

initial

Initial input value in the model when available associated with the label column

likelihood

Likelihood value for the data point

platoon

Platoon also an SS3 indexing value

age

Age of fish

bio_pattern

Indexing column of data

settlement

Indexing column of data

morph

Indexing column of data

type

Indexing column of data

factor

Indexing column of data

part

Indexing column of data

kind

Indexing column of data

nsim

Indexing column of data

bin

Indexing column of data

age_a

Alternative age column

length_bins

Length bins for composition or other length based data

count

Indexing column of data

block

Indexing column of data