Preformatted reference line

Preformatted reference line

reference_line(label_name, ref_line, scale_amount = 1, model_name = "1")

Arguments

label_name

String. Name of the quantity that users want to extract the reference point from

ref_line

String. Reference point(s)

scale_amount

Number. A number to scale the y-axis values.

Default: 1

model_name

String. Name of the model that will be present in the legend.

Options: Including, but not limited to: "msy", "unfished", "target"

Value

A ggplot2 geom_hline object for a reference point that can be added to a plot

Examples

if (FALSE) { # \dontrun{
reference_line(dat, "biomass", "msy")
} # }