Biomass, abundance, and catch time series table

Biomass, abundance, and catch time series table

table_biomass_abundance_catch(
  dat,
  unit_label = c(biomass = "mt", abundance = "fish", catch = "mt"),
  era = NULL,
  interactive = TRUE,
  group = NULL,
  method = "sum",
  module = NULL,
  label = NULL,
  make_rda = FALSE,
  tables_dir = getwd()
)

Arguments

dat

Data frame or list. A tibble or named list of tibbles (input as `list()`) returned from convert_output.

If inputting a list of tibbles, the first tibble's reference point defined in `ref_line` is used to plot a reference line or calculate relative spawning biomass.

unit_label

Character vector. Abbreviated unit label for each quantity

Default: c("biomass" = "mt", "abundance" = "fish", "catch" = "mt")

era

String. Era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)

interactive

Logical. TRUE/FALSE; indicate whether the environment is interactive.

Default: `FALSE`

group

A string of a single column that groups the data.

Set group = "none" to summarize data over all indexing values.

Default: NULL Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL

method

A string describing the method of summarizing data when group is set to "none".

Default: "sum"

Options: "sum" or "mean"

module

Character vector. (Optional) Module name found in `dat$module_name`. If selecting >1 module, place them in a vector like c("module1", "module2").

Default: NULL

If the interactive and >1 module_name is found, user will select the module_name in the console. @seealso [filter_data()]

label

The label that will be chosen from the input file. If unspecified, the function will search the "label" column and use the first matching label in this ordered list: "landings_weight", "landings_numbers", "landings_expected", "landings_predicted", "landings".

Default: NULL

make_rda

Logical. TRUE/FALSE; indicate whether to save the object and make an automated caption and alternative text in the form of an `rda` object. If TRUE, the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.

tables_dir

The location of the folder containing the table rda files ("tables") that will be created if the argument `make_rda` = TRUE. Default: the working directory (`getwd()`)

Value

A table of biomass, abundance, catch, and spawning biomass.

Details

There are options to return a [gt::gt()] object or export an rda object containing a gt-based table, caption, and LaTeX-based table.

See also

Examples

table_biomass_abundance_catch(stockplotr::example_data)
#> ℹ Processing ^biomass$
#> ! Multiple module names found in data. 
#> ℹ Environment not interactive. Selecting SPR_SERIES.
#> ℹ Processing ^abundance
#> ! Multiple module names found in data. 
#> ℹ Environment not interactive. Selecting SPR_SERIES.
#> ℹ Processing catch$|landings_observed
#> ! Multiple module names found in data. 
#> ℹ Environment not interactive. Selecting CATCH.

  

Year
      	Biomass (mt) 
      	Abundance Midyear (fish) 
      	Landings Observed Weight (mt)  (se)
    

1876
	42,640.00
	 NA
	1.00 (0.01)

1877
	42,640.00
	 NA
	1.00 (0.01)

1878
	42,640.00
	 NA
	1.00 (0.01)

1879
	42,640.00
	 NA
	1.00 (0.01)

1880
	42,552.20
	 NA
	11.55 (0.01)

1881
	42,464.70
	 NA
	22.10 (0.01)

1882
	42,377.40
	 NA
	32.65 (0.01)

1883
	42,290.30
	 NA
	43.20 (0.01)

1884
	42,203.40
	 NA
	53.75 (0.01)

1885
	42,116.60
	 NA
	64.30 (0.01)

1886
	42,029.80
	 NA
	74.85 (0.01)

1887
	41,943.10
	 NA
	85.40 (0.01)

1888
	41,856.30
	 NA
	95.95 (0.01)

1889
	41,769.60
	 NA
	106.50 (0.01)

1890
	41,682.90
	 NA
	117.05 (0.01)

1891
	41,596.10
	 NA
	127.60 (0.01)

1892
	41,509.20
	 NA
	138.15 (0.01)

1893
	41,422.20
	 NA
	148.71 (0.01)

1894
	41,335.20
	 NA
	159.26 (0.01)

1895
	41,248.10
	 NA
	169.81 (0.01)

1896
	41,159.50
	 NA
	180.60 (0.01)

1897
	41,072.50
	 NA
	191.11 (0.01)

1898
	40,985.40
	 NA
	201.61 (0.01)

1899
	40,898.00
	 NA
	212.16 (0.01)

1900
	40,810.50
	 NA
	222.71 (0.01)

1901
	40,722.80
	 NA
	233.25 (0.01)

1902
	40,635.10
	 NA
	243.80 (0.01)

1903
	40,547.20
	 NA
	254.34 (0.01)

1904
	40,459.30
	 NA
	264.89 (0.01)

1905
	40,371.30
	 NA
	275.44 (0.01)

1906
	40,283.10
	 NA
	285.98 (0.01)

1907
	40,194.80
	 NA
	296.53 (0.01)

1908
	40,106.50
	 NA
	307.07 (0.01)

1909
	40,018.00
	 NA
	317.62 (0.01)

1910
	39,929.40
	 NA
	328.16 (0.01)

1911
	39,840.80
	 NA
	338.71 (0.01)

1912
	39,752.00
	 NA
	349.26 (0.01)

1913
	39,663.10
	 NA
	359.80 (0.01)

1914
	39,574.20
	 NA
	370.35 (0.01)

1915
	39,485.10
	 NA
	380.89 (0.01)

1916
	39,433.90
	 NA
	386.50 (0.01)

1917
	38,375.50
	 NA
	526.49 (0.01)

1918
	39,120.40
	 NA
	423.92 (0.01)

1919
	39,815.70
	 NA
	333.51 (0.01)

1920
	40,649.00
	 NA
	230.55 (0.01)

1921
	40,138.80
	 NA
	293.82 (0.01)

1922
	39,115.10
	 NA
	424.83 (0.01)

1923
	39,087.40
	 NA
	427.41 (0.01)

1924
	38,288.40
	 NA
	532.91 (0.01)

1925
	38,301.30
	 NA
	528.51 (0.01)

1926
	38,334.30
	 NA
	521.71 (0.01)

1927
	37,512.90
	 NA
	632.08 (0.01)

1928
	37,569.80
	 NA
	620.09 (0.01)

1929
	36,923.40
	 NA
	707.58 (0.01)

1930
	37,223.80
	 NA
	660.05 (0.01)

1931
	37,361.90
	 NA
	641.11 (0.01)

1932
	37,697.70
	 NA
	598.39 (0.01)

1933
	38,619.00
	 NA
	478.74 (0.01)

1934
	33,857.40
	 NA
	1,179.95 (0.01)

1935
	35,001.60
	 NA
	982.24 (0.01)

1936
	37,478.00
	 NA
	626.31 (0.01)

1937
	34,893.20
	 NA
	1,015.77 (0.01)

1938
	34,032.20
	 NA
	1,123.74 (0.01)

1939
	31,902.40
	 NA
	1,534.78 (0.01)

1940
	33,346.60
	 NA
	1,316.87 (0.01)

1941
	34,397.40
	 NA
	1,176.76 (0.01)

1942
	30,442.90
	 NA
	2,206.06 (0.01)

1943
	29,337.70
	 NA
	2,375.88 (0.01)

1944
	31,555.00
	 NA
	1,621.40 (0.01)

1945
	31,858.50
	 NA
	1,508.26 (0.01)

1946
	26,241.80
	 NA
	2,805.53 (0.01)

1947
	27,185.60
	 NA
	2,222.58 (0.01)

1948
	20,655.20
	 NA
	4,280.45 (0.01)

1949
	21,775.70
	 NA
	3,069.20 (0.01)

1950
	20,427.30
	 NA
	3,461.25 (0.01)

1951
	23,414.70
	 NA
	2,323.44 (0.01)

1952
	23,290.80
	 NA
	2,176.72 (0.01)

1953
	23,493.20
	 NA
	1,928.17 (0.01)

1954
	20,664.80
	 NA
	2,525.99 (0.01)

1955
	21,217.10
	 NA
	2,294.32 (0.01)

1956
	23,058.80
	 NA
	1,849.25 (0.01)

1957
	19,771.00
	 NA
	2,639.74 (0.01)

1958
	20,401.80
	 NA
	2,318.67 (0.01)

1959
	21,995.00
	 NA
	1,880.98 (0.01)

1960
	20,955.30
	 NA
	2,157.46 (0.01)

1961
	18,754.20
	 NA
	2,548.42 (0.01)

1962
	18,059.70
	 NA
	2,690.84 (0.01)

1963
	17,227.20
	 NA
	2,710.56 (0.01)

1964
	18,882.30
	 NA
	2,150.66 (0.01)

1965
	19,258.70
	 NA
	2,109.23 (0.01)

1966
	18,833.80
	 NA
	2,289.51 (0.01)

1967
	19,522.20
	 NA
	2,181.42 (0.01)

1968
	19,929.80
	 NA
	2,121.13 (0.01)

1969
	19,942.20
	 NA
	2,249.00 (0.01)

1970
	18,887.50
	 NA
	2,713.96 (0.01)

1971
	18,916.60
	 NA
	2,887.68 (0.01)

1972
	19,015.60
	 NA
	2,953.35 (0.01)

1973
	17,929.20
	 NA
	2,950.56 (0.01)

1974
	15,433.30
	 NA
	3,624.83 (0.01)

1975
	14,514.30
	 NA
	3,489.85 (0.01)

1976
	14,709.70
	 NA
	2,764.06 (0.01)

1977
	15,567.60
	 NA
	2,267.94 (0.01)

1978
	12,231.30
	 NA
	3,058.79 (0.01)

1979
	10,294.80
	 NA
	3,054.97 (0.01)

1980
	10,640.70
	 NA
	2,467.60 (0.01)

1981
	11,457.60
	 NA
	2,041.19 (0.01)

1982
	9,727.88
	 NA
	2,630.33 (0.01)

1983
	10,589.50
	 NA
	2,214.19 (0.01)

1984
	11,045.40
	 NA
	1,739.17 (0.01)

1985
	10,124.50
	 NA
	1,839.45 (0.01)

1986
	10,130.70
	 NA
	1,750.43 (0.01)

1987
	8,739.28
	 NA
	2,204.84 (0.01)

1988
	8,668.73
	 NA
	2,149.18 (0.01)

1989
	8,701.67
	 NA
	2,152.70 (0.01)

1990
	9,644.76
	 NA
	1,764.64 (0.01)

1991
	8,967.17
	 NA
	1,927.37 (0.01)

1992
	9,743.79
	 NA
	1,553.50 (0.01)

1993
	10,082.60
	 NA
	1,503.21 (0.01)

1994
	11,372.50
	 NA
	1,375.14 (0.01)

1995
	11,084.70
	 NA
	1,659.05 (0.01)

1996
	10,863.00
	 NA
	1,828.60 (0.01)

1997
	10,502.20
	 NA
	1,947.72 (0.01)

1998
	12,509.10
	 NA
	1,462.77 (0.01)

1999
	13,228.50
	 NA
	1,497.11 (0.01)

2000
	12,248.10
	 NA
	1,893.27 (0.01)

2001
	12,603.40
	 NA
	1,845.22 (0.01)

2002
	12,868.00
	 NA
	1,797.09 (0.01)

2003
	12,995.50
	 NA
	2,069.58 (0.01)

2004
	14,009.30
	 NA
	1,963.97 (0.01)

2005
	12,226.60
	 NA
	2,734.36 (0.01)

2006
	12,117.30
	 NA
	2,610.20 (0.01)

2007
	12,477.90
	 NA
	2,252.84 (0.01)

2008
	12,125.30
	 NA
	2,219.71 (0.01)

2009
	13,032.60
	 NA
	1,767.44 (0.01)

2010
	19,752.60
	 NA
	803.10 (0.01)

2011
	22,244.50
	 NA
	934.51 (0.01)

2012
	23,703.30
	 NA
	1,117.72 (0.01)

2013
	20,093.40
	 NA
	2,253.36 (0.01)

2014
	20,976.70
	 NA
	2,408.74 (0.01)

2015
	20,958.60
	 NA
	2,665.17 (0.01)

2016
	21,129.90
	 NA
	2,727.63 (0.01)

2017
	20,520.40
	 NA
	2,930.62 (0.01)

2018
	20,831.50
	 NA
	2,894.44 (0.01)

2019
	21,423.40
	 NA
	2,616.91 (0.01)

2020
	23,125.50
	 NA
	2,092.13 (0.01)

2021
	20,253.80
	 NA
	2,879.11 (0.01)

2022
	19,064.90
	 NA
	3,059.94 (0.01)

table_biomass_abundance_catch(stockplotr::example_data,
  unit_label = c("biomass" = "kg", "abundance" = "eggs", "catch" = "kg"),
  module = c("TIME_SERIES", "CATCH", "TIME_SERIES"),
  interactive = FALSE
)
#> ℹ Processing ^biomass$
#> ℹ Processing ^abundance
#> Warning: Values from `Uncertainty` and `estimate` are not uniquely identified; output
#> will contain list-cols.
#> • Use `values_fn = list` to suppress this warning.
#> • Use `values_fn = {summary_fun}` to summarise duplicates.
#> • Use the following dplyr code to identify duplicates.
#>   {data} |>
#>   dplyr::summarise(n = dplyr::n(), .by = c(Year, label)) |>
#>   dplyr::filter(n > 1L)
#> ℹ Processing catch$|landings_observed

  

Year
      	Biomass (kg) 
      	Abundance Midyear (eggs) 
      	Landings Observed Weight (kg)  (se)
    

1876
	42,648.70
	c("37,951.60", "33,655.40") (c(NA, NA))
	1.00 (0.01)

1877
	42,647.80
	c("37,951.10", "33,654.90") (c(NA, NA))
	1.00 (0.01)

1878
	42,647.10
	c("37,950.70", "33,654.50") (c(NA, NA))
	1.00 (0.01)

1879
	42,646.60
	c("37,951.70", "33,655.50") (c(NA, NA))
	1.00 (0.01)

1880
	42,646.40
	c("37,952.50", "33,656.30") (c(NA, NA))
	11.55 (0.01)

1881
	42,636.70
	c("37,947.40", "33,651.50") (c(NA, NA))
	22.10 (0.01)

1882
	42,618.30
	c("37,937.10", "33,641.90") (c(NA, NA))
	32.65 (0.01)

1883
	42,591.90
	c("37,922.20", "33,628.10") (c(NA, NA))
	43.20 (0.01)

1884
	42,558.20
	c("37,903.30", "33,610.60") (c(NA, NA))
	53.75 (0.01)

1885
	42,517.80
	c("37,880.90", "33,590.00") (c(NA, NA))
	64.30 (0.01)

1886
	42,471.60
	c("37,855.30", "33,566.70") (c(NA, NA))
	74.85 (0.01)

1887
	42,419.90
	c("37,826.90", "33,540.80") (c(NA, NA))
	85.40 (0.01)

1888
	42,363.50
	c("37,796.00", "33,512.80") (c(NA, NA))
	95.95 (0.01)

1889
	42,302.80
	c("37,762.90", "33,483.00") (c(NA, NA))
	106.50 (0.01)

1890
	42,238.30
	c("37,727.80", "33,451.40") (c(NA, NA))
	117.05 (0.01)

1891
	42,170.30
	c("37,691.00", "33,418.40") (c(NA, NA))
	127.60 (0.01)

1892
	42,099.40
	c("37,652.60", "33,384.00") (c(NA, NA))
	138.15 (0.01)

1893
	42,025.70
	c("37,612.70", "33,348.50") (c(NA, NA))
	148.71 (0.01)

1894
	41,949.60
	c("37,571.60", "33,312.00") (c(NA, NA))
	159.26 (0.01)

1895
	41,871.40
	c("37,529.40", "33,274.60") (c(NA, NA))
	169.81 (0.01)

1896
	41,791.20
	c("37,486.20", "33,236.30") (c(NA, NA))
	180.60 (0.01)

1897
	41,709.20
	c("37,442.00", "33,197.30") (c(NA, NA))
	191.11 (0.01)

1898
	41,625.70
	c("37,397.00", "33,157.60") (c(NA, NA))
	201.61 (0.01)

1899
	41,540.90
	c("37,351.30", "33,117.40") (c(NA, NA))
	212.16 (0.01)

1900
	41,454.90
	c("37,305.00", "33,076.80") (c(NA, NA))
	222.71 (0.01)

1901
	41,367.90
	c("37,258.10", "33,035.70") (c(NA, NA))
	233.25 (0.01)

1902
	41,279.90
	c("37,210.80", "32,994.20") (c(NA, NA))
	243.80 (0.01)

1903
	41,191.10
	c("37,163.00", "32,952.50") (c(NA, NA))
	254.34 (0.01)

1904
	41,101.50
	c("37,114.80", "32,910.40") (c(NA, NA))
	264.89 (0.01)

1905
	41,011.30
	c("37,066.30", "32,868.20") (c(NA, NA))
	275.44 (0.01)

1906
	40,920.50
	c("37,017.50", "32,825.70") (c(NA, NA))
	285.98 (0.01)

1907
	40,829.20
	c("36,968.50", "32,783.00") (c(NA, NA))
	296.53 (0.01)

1908
	40,737.50
	c("36,919.30", "32,740.30") (c(NA, NA))
	307.07 (0.01)

1909
	40,645.40
	c("36,869.90", "32,697.40") (c(NA, NA))
	317.62 (0.01)

1910
	40,553.00
	c("36,820.50", "32,654.60") (c(NA, NA))
	328.16 (0.01)

1911
	40,460.30
	c("36,770.90", "32,611.60") (c(NA, NA))
	338.71 (0.01)

1912
	40,367.40
	c("36,721.40", "32,568.70") (c(NA, NA))
	349.26 (0.01)

1913
	40,274.30
	c("36,671.80", "32,525.90") (c(NA, NA))
	359.80 (0.01)

1914
	40,181.20
	c("36,622.40", "32,483.20") (c(NA, NA))
	370.35 (0.01)

1915
	40,088.00
	c("36,573.00", "32,440.60") (c(NA, NA))
	380.89 (0.01)

1916
	39,994.70
	c("36,523.80", "32,398.20") (c(NA, NA))
	386.50 (0.01)

1917
	39,906.10
	c("36,477.60", "32,358.60") (c(NA, NA))
	526.49 (0.01)

1918
	39,698.50
	c("36,357.70", "32,249.90") (c(NA, NA))
	423.92 (0.01)

1919
	39,603.30
	c("36,311.40", "32,211.40") (c(NA, NA))
	333.51 (0.01)

1920
	39,601.90
	c("36,323.00", "32,226.70") (c(NA, NA))
	230.55 (0.01)

1921
	39,698.80
	c("36,391.10", "32,293.90") (c(NA, NA))
	293.82 (0.01)

1922
	39,733.30
	c("36,414.90", "32,318.00") (c(NA, NA))
	424.83 (0.01)

1923
	39,646.60
	c("36,362.10", "32,269.90") (c(NA, NA))
	427.41 (0.01)

1924
	39,565.00
	c("36,317.10", "32,229.80") (c(NA, NA))
	532.91 (0.01)

1925
	39,394.60
	c("36,219.90", "32,141.30") (c(NA, NA))
	528.51 (0.01)

1926
	39,243.40
	c("36,139.00", "32,069.00") (c(NA, NA))
	521.71 (0.01)

1927
	39,113.40
	c("36,074.10", "32,012.30") (c(NA, NA))
	632.08 (0.01)

1928
	38,896.40
	c("35,955.60", "31,905.70") (c(NA, NA))
	620.09 (0.01)

1929
	38,711.50
	c("35,860.70", "31,822.20") (c(NA, NA))
	707.58 (0.01)

1930
	38,466.30
	c("35,730.50", "31,706.10") (c(NA, NA))
	660.05 (0.01)

1931
	38,289.70
	c("35,646.60", "31,634.30") (c(NA, NA))
	641.11 (0.01)

1932
	38,153.60
	c("35,591.50", "31,587.30") (c(NA, NA))
	598.39 (0.01)

1933
	38,079.00
	c("35,578.10", "31,575.90") (c(NA, NA))
	478.74 (0.01)

1934
	38,131.50
	c("35,645.70", "31,638.40") (c(NA, NA))
	1,179.95 (0.01)

1935
	37,549.60
	c("35,312.00", "31,324.30") (c(NA, NA))
	982.24 (0.01)

1936
	37,203.90
	c("35,151.00", "31,175.80") (c(NA, NA))
	626.31 (0.01)

1937
	37,229.20
	c("35,241.30", "31,259.20") (c(NA, NA))
	1,015.77 (0.01)

1938
	36,923.00
	c("35,118.70", "31,132.30") (c(NA, NA))
	1,123.74 (0.01)

1939
	36,543.90
	c("34,927.10", "30,954.50") (c(NA, NA))
	1,534.78 (0.01)

1940
	35,846.20
	c("34,536.00", "30,562.80") (c(NA, NA))
	1,316.87 (0.01)

1941
	35,417.40
	c("34,285.60", "30,278.20") (c(NA, NA))
	1,176.76 (0.01)

1942
	35,141.60
	c("34,029.40", "29,970.00") (c(NA, NA))
	2,206.06 (0.01)

1943
	34,030.70
	c("33,206.30", "28,996.80") (c(NA, NA))
	2,375.88 (0.01)

1944
	32,781.00
	c("32,114.30", "27,817.10") (c(NA, NA))
	1,621.40 (0.01)

1945
	32,118.30
	c("31,197.90", "26,936.70") (c(NA, NA))
	1,508.26 (0.01)

1946
	31,491.40
	c("30,325.40", "26,124.40") (c(NA, NA))
	2,805.53 (0.01)

1947
	29,725.60
	c("28,988.20", "24,844.50") (c(NA, NA))
	2,222.58 (0.01)

1948
	28,492.80
	c("28,206.30", "24,202.20") (c(NA, NA))
	4,280.45 (0.01)

1949
	25,580.20
	c("26,497.60", "22,651.30") (c(NA, NA))
	3,069.20 (0.01)

1950
	23,835.70
	c("25,364.30", "21,791.30") (c(NA, NA))
	3,461.25 (0.01)

1951
	21,938.50
	c("24,169.60", "20,805.50") (c(NA, NA))
	2,323.44 (0.01)

1952
	21,187.10
	c("23,714.60", "20,496.00") (c(NA, NA))
	2,176.72 (0.01)

1953
	20,623.40
	c("23,394.50", "20,315.10") (c(NA, NA))
	1,928.17 (0.01)

1954
	20,284.40
	c("23,188.00", "20,266.00") (c(NA, NA))
	2,525.99 (0.01)

1955
	19,443.80
	c("22,590.60", "19,824.80") (c(NA, NA))
	2,294.32 (0.01)

1956
	18,854.20
	c("22,110.60", "19,461.90") (c(NA, NA))
	1,849.25 (0.01)

1957
	18,675.90
	c("21,900.80", "19,322.50") (c(NA, NA))
	2,639.74 (0.01)

1958
	17,808.60
	c("21,145.00", "18,624.00") (c(NA, NA))
	2,318.67 (0.01)

1959
	17,221.60
	c("20,464.20", "18,009.70") (c(NA, NA))
	1,880.98 (0.01)

1960
	17,042.80
	c("20,101.20", "17,700.20") (c(NA, NA))
	2,157.46 (0.01)

1961
	16,785.30
	c("20,111.60", "17,709.20") (c(NA, NA))
	2,548.42 (0.01)

1962
	16,423.30
	c("21,214.30", "18,828.10") (c(NA, NA))
	2,690.84 (0.01)

1963
	16,206.00
	c("23,329.80", "20,915.70") (c(NA, NA))
	2,710.56 (0.01)

1964
	16,065.30
	c("23,035.60", "20,716.90") (c(NA, NA))
	2,150.66 (0.01)

1965
	16,412.80
	c("22,368.30", "20,117.10") (c(NA, NA))
	2,109.23 (0.01)

1966
	16,826.30
	c("22,171.50", "19,918.30") (c(NA, NA))
	2,289.51 (0.01)

1967
	17,471.20
	c("21,760.50", "19,474.60") (c(NA, NA))
	2,181.42 (0.01)

1968
	18,631.40
	c("30,307.70", "27,705.10") (c(NA, NA))
	2,121.13 (0.01)

1969
	19,892.30
	c("31,447.60", "28,771.70") (c(NA, NA))
	2,249.00 (0.01)

1970
	20,742.10
	c("28,643.70", "26,002.30") (c(NA, NA))
	2,713.96 (0.01)

1971
	20,690.20
	c("25,340.10", "22,707.60") (c(NA, NA))
	2,887.68 (0.01)

1972
	20,001.20
	c("22,526.00", "19,827.70") (c(NA, NA))
	2,953.35 (0.01)

1973
	18,873.40
	c("20,589.40", "17,811.40") (c(NA, NA))
	2,950.56 (0.01)

1974
	17,414.00
	c("19,224.70", "16,504.50") (c(NA, NA))
	3,624.83 (0.01)

1975
	15,219.80
	c("17,233.00", "14,680.40") (c(NA, NA))
	3,489.85 (0.01)

1976
	13,144.30
	c("15,370.90", "13,079.40") (c(NA, NA))
	2,764.06 (0.01)

1977
	11,806.70
	c("14,545.00", "12,526.90") (c(NA, NA))
	2,267.94 (0.01)

1978
	11,132.40
	c("14,888.10", "13,079.00") (c(NA, NA))
	3,058.79 (0.01)

1979
	9,911.76
	c("15,349.70", "13,802.50") (c(NA, NA))
	3,054.97 (0.01)

1980
	8,660.89
	c("13,796.70", "12,566.50") (c(NA, NA))
	2,467.60 (0.01)

1981
	7,904.04
	c("11,875.00", "10,845.20") (c(NA, NA))
	2,041.19 (0.01)

1982
	7,514.75
	c("11,358.90", "10,416.50") (c(NA, NA))
	2,630.33 (0.01)

1983
	6,618.73
	c("10,831.80", "9,988.21") (c(NA, NA))
	2,214.19 (0.01)

1984
	6,104.36
	c("10,524.00", "9,123.98") (c(NA, NA))
	1,739.17 (0.01)

1985
	6,109.04
	c("10,579.00", "8,909.95") (c(NA, NA))
	1,839.45 (0.01)

1986
	6,178.65
	c("12,188.70", "10,247.00") (c(NA, NA))
	1,750.43 (0.01)

1987
	6,425.57
	c("13,588.90", "11,487.80") (c(NA, NA))
	2,204.84 (0.01)

1988
	6,212.39
	c("12,666.20", "10,480.40") (c(NA, NA))
	2,149.18 (0.01)

1989
	5,912.57
	c("10,818.10", "8,553.03") (c(NA, NA))
	2,152.70 (0.01)

1990
	5,528.16
	c("10,075.50", "7,736.32") (c(NA, NA))
	1,764.64 (0.01)

1991
	5,504.11
	c("10,642.00", "8,372.66") (c(NA, NA))
	1,927.37 (0.01)

1992
	5,420.45
	c("11,235.80", "9,135.75") (c(NA, NA))
	1,553.50 (0.01)

1993
	5,723.00
	c("11,996.70", "10,005.90") (c(NA, NA))
	1,503.21 (0.01)

1994
	6,053.10
	c("12,020.40", "9,970.69") (c(NA, NA))
	1,375.14 (0.01)

1995
	6,533.96
	c("11,413.20", "9,212.90") (c(NA, NA))
	1,659.05 (0.01)

1996
	6,819.65
	c("13,148.40", "10,723.70") (c(NA, NA))
	1,828.60 (0.01)

1997
	6,986.83
	c("13,254.20", "10,714.00") (c(NA, NA))
	1,947.72 (0.01)

1998
	7,039.81
	c("13,411.20", "10,837.80") (c(NA, NA))
	1,462.77 (0.01)

1999
	7,537.44
	c("13,289.30", "10,728.10") (c(NA, NA))
	1,497.11 (0.01)

2000
	8,163.01
	c("13,594.10", "10,946.50") (c(NA, NA))
	1,893.27 (0.01)

2001
	8,608.66
	c("16,761.80", "13,917.20") (c(NA, NA))
	1,845.22 (0.01)

2002
	9,146.03
	c("17,341.80", "14,448.90") (c(NA, NA))
	1,797.09 (0.01)

2003
	9,589.10
	c("16,684.40", "13,763.30") (c(NA, NA))
	2,069.58 (0.01)

2004
	9,746.01
	c("15,837.40", "12,905.10") (c(NA, NA))
	1,963.97 (0.01)

2005
	9,849.40
	c("15,323.10", "12,297.50") (c(NA, NA))
	2,734.36 (0.01)

2006
	9,201.90
	c("14,275.60", "11,165.10") (c(NA, NA))
	2,610.20 (0.01)

2007
	8,753.74
	c("14,090.70", "11,069.00") (c(NA, NA))
	2,252.84 (0.01)

2008
	9,040.86
	c("15,199.10", "12,275.00") (c(NA, NA))
	2,219.71 (0.01)

2009
	10,012.90
	c("20,636.10", "17,638.20") (c(NA, NA))
	1,767.44 (0.01)

2010
	11,949.10
	c("25,277.60", "22,174.20") (c(NA, NA))
	803.10 (0.01)

2011
	15,040.00
	c("30,755.50", "27,384.50") (c(NA, NA))
	934.51 (0.01)

2012
	17,973.90
	c("30,200.90", "26,633.80") (c(NA, NA))
	1,117.72 (0.01)

2013
	20,346.80
	c("28,845.10", "25,023.30") (c(NA, NA))
	2,253.36 (0.01)

2014
	21,348.60
	c("27,414.10", "23,039.60") (c(NA, NA))
	2,408.74 (0.01)

2015
	21,799.40
	c("28,812.60", "23,924.70") (c(NA, NA))
	2,665.17 (0.01)

2016
	21,705.90
	c("26,604.10", "21,500.60") (c(NA, NA))
	2,727.63 (0.01)

2017
	21,340.20
	c("26,115.10", "20,937.70") (c(NA, NA))
	2,930.62 (0.01)

2018
	20,599.90
	c("25,616.30", "20,455.40") (c(NA, NA))
	2,894.44 (0.01)

2019
	19,738.60
	c("23,831.70", "18,746.10") (c(NA, NA))
	2,616.91 (0.01)

2020
	18,950.00
	c("22,093.30", "17,172.30") (c(NA, NA))
	2,092.13 (0.01)

2021
	18,500.90
	c("21,198.60", "16,466.10") (c(NA, NA))
	2,879.11 (0.01)

2022
	17,372.50
	c("20,613.30", "15,991.80") (c(NA, NA))
	3,059.94 (0.01)